pyproject.toml
[project]
name = "blockchainproj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "pexpect>=4.9.0",
    "python-bitcoinlib>=0.12.2",
]
